+++ src/index-es.js
@@ -1,4 +1,4 @@
-import Wfst, { WfsLayer, WmsLayer, Geoserver } from '../src/ol-wfst';
+import Wfst from './ol-wfst';
 
 export default Wfst;
-export { WfsLayer, WmsLayer, Geoserver };
+export * from './ol-wfst';